backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
python-dotenv==1.0.1
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
anyio==4.6.2
email-validator==2.2.0
pydantic-settings==2.6.1
httpx==0.27.0
authlib==1.3.1
